packages/dali_pipeline_framework/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel", "torch>=2.0.0", "scikit-build>=0.17.0", "pybind11>=2.10.0"]
build-backend = "setuptools.build_meta"

[project]
name = "accvlab.dali_pipeline_framework"
version = "0.1.0"
description = "DALI Pipeline Framework Package (part of the ACCV-Lab package)."
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.22.2",
    "accvlab-build-config>=0.1.0",
    "nvidia-dali-cuda120>=1.51.2",
    "numba>=0.59",
]

[project.optional-dependencies]
optional = [
    "pytest",
    "opencv-python-headless",
    "cupy",
    "pyquaternion",
    "nuscenes-devkit",
    "shapely",
    "psutil",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["accvlab.dali_pipeline_framework*"]

[tool.skbuild]
cmake.source-dir = "ext_impl"
cmake.minimum-version = "3.18"
cmake.install-dir = "." 